backend/requirements.txt
django==4.1.2
djangorestframework==3.14.0
python-slugify==6.1.2
requests==2.28.1
Pillow==9.2.0
selenium==4.5.0
